Reorganising project
