+++ src/utils/api.ts
@@ -1,0 +1,17 @@
+const resolveDid = async (did: string) => {
+  const res = await fetch(
+    did.startsWith("did:web") ?
+      `https://${did.split(":")[2]}/.well-known/did.json`
+    : "https://plc.directory/" + did,
+  );
+
+  return res.json().then((doc) => {
+    for (const alias of doc.alsoKnownAs) {
+      if (alias.includes("at://")) {
+        return alias.split("//")[1];
+      }
+    }
+  });
+};
+
+export default resolveDid;
