+++ src/utils/constants.ts
@@ -1,0 +1,6 @@
+const CHARLIMIT = 64;
+const MAXPOSTS = 100;
+const SERVER_URL = "pico.api.bsky.mom";
+const APP_NAME = "picosky"
+
+export { CHARLIMIT, MAXPOSTS, SERVER_URL, APP_NAME };
