+++ src/utils/defs.ts
@@ -25,3 +25,12 @@ declare module "@atcute/client/lexicons" {
 
   interface Procedures {}
 }
+
+interface PostRecord {
+  uri: string;
+  post: string;
+  handle: string;
+  indexedAt: number;
+};
+
+export type { PostRecord };
